No perf impact on FindAccessToken logic...

Before
|                         Method | TokenCacheSize |     Mean |   Error |  StdDev |
|------------------------------- |--------------- |---------:|--------:|--------:|
| AcquireTokenForClientTestAsync |         100000 | 104.9 ms | 2.06 ms | 2.37 ms |

After
|                         Method | TokenCacheSize |     Mean |   Error |  StdDev |
|------------------------------- |--------------- |---------:|--------:|--------:|
| AcquireTokenForClientTestAsync |         100000 | 104.4 ms | 2.05 ms | 2.52 ms |